app/utils/security.py
from datetime import datetime, timedelta
from jose import jwt
from passlib.hash import bcrypt
from ..config import settings

def create_access_token(sub: str, expires_minutes: int = 60):
    payload = {
        "sub": sub,
        "exp": datetime.utcnow() + timedelta(minutes=expires_minutes),
        "iat": datetime.utcnow(),
    }
    return jwt.encode(payload, settings.JWT_SECRET, algorithm=settings.JWT_ALG)

def hash_password(password: str) -> str:
    return bcrypt.hash(password)

def verify_password(password: str, password_hash: str) -> bool:
    return bcrypt.verify(password, password_hash)
